Create the xml scheme utility.

+++ src/com/moka/components/Bullet.java
@@ -1,10 +1,12 @@
 package com.moka.components;
 
+import com.moka.core.Component;
 import com.moka.core.xml.XmlAttribute;
 import com.moka.math.Vector2f;
 import com.moka.physics.Collision;
 
-public class Bullet extends Component {
+public class Bullet extends Component
+{
 	private Vector2f buffer = new Vector2f();
 	private float speed = 300;
 
@@ -33,7 +35,7 @@ public class Bullet extends Component {
 	}
 
 	@XmlAttribute("speed")
-	public void setSpeed(float speed)
+	public void setSpeed(int speed)
 	{
 		this.speed = speed;
 	}
